Implement pagination for the "variables" request
Summary:
# Context
The "variables" request is used to fetch variables and also subparts of a variable (e.g., all associations of a map).

# Problem
The DAP specification allows for pagination in the "variables" request when both the client and server support it. This makes sense when inspecting very large objects, as otherwise the cost of serialization, transmission and deserialization would make things super slow.

# This diff
We add support for pagination to the request. The idea is:

* If the client declares that it supports pagination, and the term has structure, then we include a `indexedVariables` field in the responses for `scopes` and `variables`.
* The client can ue this value to decide if it wants to paginate or not. For e.g., VSCode uses pages of 100 elements
* If the client wants to paginate, it will include start `start` and(/or?) `count`  fields in the `variables` argument. In that case, we return just the requested data

The DAP specification also mentions as `namedVariables` that can co-exist with `indexedVariables` and a filter in the `variables` request to to specify `indexed` variables, `named` variables or both. The docs are not clear about the distinction, and from trying it out, VS Code seems to ignore namedVariables` altogether, so for now we are not supporting any of that.

Reviewed By: robertoaloi

Differential Revision: D74737927

fbshipit-source-id: f63de61e852b2d6e864cc7f8abae9aa103280cbd