Implement the "evaluate" request
Summary:
# Context
The "evaluate" request is used to drive things like "watch" expressions, "repl" evaluation, copying a value to the clipboard, etc.

# Problem
At the moment we don't support this request

# This diff
Add an implementation for the request. At the moment, we don't consider what is the "context" for the evaluation, and return the same thing all the time (we could format it differently for hover, as a follow up)

Reviewed By: robertoaloi

Differential Revision: D74896986

fbshipit-source-id: 909fd05a76582c3a49ee1ebaa1c035f13b372c3c