Fix spec of `edb:stack_frame_vars()`
Summary:
# Context
The `edb:stack_frame_vars()` function is used to get the variables on some stack-frame of a process. The `MaxTermSize` argument, controls what's the largest size of a term one is willing to receive.

# Problem
Literals (atoms, small ints, etc) fit in a word, so they don't take heap space. Their size is 0. However, the spec of the function currently asks for a value of at least 1.

Under the hood, this function calls functoins in `erl_debugger`, where max term sizes are specified as `non_neg_integer()`

# This diff
Use `non_neg_integer()`

Reviewed By: TD5

Differential Revision: D74934304

fbshipit-source-id: 9c58576c57d81a730f5897519996875f1005ff78