Include "evaluateName" in "variables" response
Summary:
# Context
When a value is displayed, say, in a "variables" or "watch" context, we already offer to expand the structure of the term (for maps, lists, etc), and support pagination in case the term is gigantic.

However, we are still rendering the value of the term in full as a `binary()` for the sake of displaying. This defeats the purpose of pagination, as we'll already blow-up memory/timeout trying to serialize and transfer the full representation

We would like to instead shorten the representation and use ellipses, but that means that trying to copy to the clipboard the value of a variable that was shorten, would not work well, as we'd get the shortened version, that loses information.

To make this work, The DAP specification has the concept of a `clipboard` context for the "evaluate" request. So that the client can make such a request to get the value to include in the clipboard, and the server can return the full representation in that case.

# Problem
For this to work, in the response of the "variables" request, we need to include, for each value, an `evaluateName` field that the client needs to use, when calling `evalute` to evaluate that same value.

For scope variables this is trivially the variable name, but once we want to use it to copy a substructure, this becomes a bit trickier, as we need to start including things like `map:get(some_key, VarName)`

# This diff
Here we make it so that every "variable" response from the "variables" request includes a `evaluateName` field, as long as the value has structure. Otherwise, the client can just take the value to put in the clipboard from the returned value.

Notice that atm we can't "evaluate" registers, so that part goes unsupported for now. We could provide a special pseudo-expression in the future to evaluate those as well.

Reviewed By: TD5

Differential Revision: D74936975

fbshipit-source-id: 0a247a97a14e88b8b0fb4455d5a2d76496eeb35f