Allow for module dependencies in edb_dap_eval_delegate
Summary:
# Context
The `edb_dap_eval_delegate` is the one module from the DAP server that we use for running code on the debuggee node. We have some safeguards to try to ensure we don't introduce accidental dependencies, as the edb_dap_delegate is the only module uploaded to the debuggee.

# Problem
For evaluating expressions, we will want to also upload the ad-hoc module created for the expression. We have added a `dependencies` option to `edb:eval/1`, but it is not used by `edb_dap_eval_delegate`

# This diff
We make it so that a `edb_dap_eval_delegate:callback()` now contains a list of module dependencies that we pass to `edb:eval/1`.

Right now, no callback uses it, but we'll start using it next

Reviewed By: TD5

Differential Revision: D74897022

fbshipit-source-id: ed8c6d500ad5c7b867e35dec753d6ec07c36d037