Limit the size of value representations
Summary:
# Context
When a value is displayed, say, in a "variables" or "watch" context, we already offer to expand the structure of the term (for maps, lists, etc), and support pagination in case the term is gigantic.

# Problem
However, we are still rendering the value of the term in full as a `binary()` for the sake of displaying. This defeats the purpose of pagination, as we'll already blow-up memory/timeout trying to serialize and transfer the full representation.

# This diff
We use `~P` to limit the depth when rendering the representation of the value. In order for "copying-to-clipboard" to still work, we implement a special `clipboard` evaluation context that will not truncate the representatio, and inform the client we support this capability

Reviewed By: TD5

Differential Revision: D74936974

fbshipit-source-id: d7bc34e3ba3923aae3911f7fbd3af22588de3e02