Add support for evaluating erlang expressions
Summary:
# Context
So far, we have a added support to evaluate function refs in the context of a process frame.

# Problem
In order to support watch expressions, conditional breakpoints, etc, we need to let the user provide strings representing Erlang expressions, and dynamically compile them

# This diff
We add a `edb_expr` module with `compile_expr/2` and `compile_guard/2` functions, that dynamically compile expressions, given their promised free-variables.

The way this works:
  - For each expression, we will generate a new content-addressable module and load it.
    - Therefore, we can reuse a previously compiled module for the same expression, just be checking if it is loaded
 - We also stash it in a persistent term, since `code:get_object_code/1` only works for modules that can be found in the path
 - We ensure that runtime and compilation errors can only happen on user provided code, and that the reported locations are correct relative to the start of the given locations.

Notice that we put this in a separate `edb_expr` and won't expose specific functionality in `edb` itself. This way, we commit to a more language agnostic API, and ensure that other Beam languages can provide their own expression compilers that can be plugged to the main API.

Reviewed By: TD5

Differential Revision: D74897021

fbshipit-source-id: b1880204a7523ad90833ddef77cdf32f9100882d