Allow explicit module dependencies in edb:eval/1
Summary:
# Context
The `edb:eval/1` function let's the caller execute a function reference on the debuggee node, in the context of a process frame.

At the moment the implementation ensures that the module of the function ref is uploaded, if necessary, to the debuggee node, but makes the caller's responsibility to ensure any further dependency is uploaded.

# Problem
It is inconvenient for the caller to transfer modules and the logic for `edb:eval/1` already handles that.

# This diff
We add an optional `dependencies` option, with a list of modules, so that the caller can declare the necessary dependencies and the `edb:eval/1` will upload them as well.

Reviewed By: TD5

Differential Revision: D74897018

fbshipit-source-id: c2e6904ac3ae61482d961ab1f29d95a1004d4440