Support reverse attach on nodes started with `-sname undefined`
Summary:
# Context
When reverse-attaching to a new node, we inject some code that the node needs to execute to register with the debugger node.

# Problem
When a node is started with `-sname undefined`, the net kernel is started but the node is not given a name until it actually connects to some other node.

The code being injected tries to do `gen_server:call()` to the debugger node before properly connecting, and this fails

This is annoying since for running tests, etc, it is very convenient to get a name dynamically generated

# This diff
We add a `net_kernel:connect_node(...)` to the code we inject on the debuggee, so that it works in all cases.

Reviewed By: robertoaloi

Differential Revision: D75138584

fbshipit-source-id: fb97196d20a2dc9ae6d43e77864398d88703f7b4