Support reverse-attach on nodes with no distribution
Summary:
# Context
When reverse-attaching to a new node, we inject some code that the node needs to execute to register with the debugger node.

# Problem
At the moment, the debuggee node had to be started with distribution, by either providing the `-name` or `-sname` options.

# This diff
We make it so that reverse-attaching even if this was not the case.

We will now detect whether the node has distribution started, and start it, if necessary

Reviewed By: robertoaloi

Differential Revision: D75140476

fbshipit-source-id: 02474d55168b2c2c898922cfa68fb2cc5dfdafc5