Add Registered name info to the "Process scope"
Summary:
# Context

When a process is paused, we can see, a part from the "Local" variables, some "Process" information. At the moment we only show messages in the process mailbox.

# Problem

There's more information that could help understand a process' state better, including the registered name of the process.

# This diff

Extends ```edb_dap_eval_delegate:process_scope()``` to also present the registered name of a process.

Reviewed By: jcpetruzza

Differential Revision: D77657717

fbshipit-source-id: e33cae0d10b90e823e857b6f5da0ff22d973dcb9