Improve handling of free-variables when evaluating
Summary:
# Context
To evaluate an expression we currently do the following:
  1. Create a content-addressable module with a function to evaluate the expression
  2. Use `edb:eval()` to evaluate this function

Because the module's name is content-addressable, we avoid recompilation/reloading.

The function to evaluate matches against the free-variables

# Problem
We have two issues related to free-variables:

1. Currently we only consider as "free variables" only those passed by the caller, that correspond to the known variables of the processes at the point of evaluation. Because of this, even for a ground term, the content-address will be likely different in each invocation, so we'll typically get cache-misses
2. For frames where we don't have variable information, we are still expecting to an empty list of variables, which doesn't correspond to what `edb:eval()` does in that case

# This diff
We now use the syntaxTools library to infer the free variables in the expression and take the intersection with the variables in scope. In particular, for ground terms, we'll always assume no free-variables, maximizing caching.

Furthermore, when there are no free-variables, we now don't expect a `vars` entry to be given in the input to the evaluation function, so that evaluating ground terms no works in every frame (before, it would raise in those without variable information)

Reviewed By: robertoaloi

Differential Revision: D77803564

fbshipit-source-id: e36d95c17348f839f3e76bc4e519bc2f2e2af4c5