Remove usages of deprecated `catch` statement
Summary:
# Context
The `catch` statement is deprecated and expected to go away from the language.

# Problem
We have a couple of usages in edb, mostly when we want to stop a node and don't care about any exceptions.

# This diff
Use a try-catch instead in those cases.

Notice that we still have usages of `catch` in tests, but mostly to test the interaction of edb with code using `catch`. Those can be removed in the future, after `catch` is gone

Reviewed By: TD5

Differential Revision: D77927748

fbshipit-source-id: 8261506b8aeff40c81646b12860955d4897d9ed4