Avoid hardwired pid in process mailbox test
Summary:
# Context
In the "Process" scope, we show the process mailbox, by default just the count and if the user expands it or copies it to the clipboard, they get a list of messages.

# Problem
We have a unit case for it and we like to be able to match the expected response by equality. However, the "evaluateName" expression, that is needed e.g. to be able to get the value to copy to the clipboard, is an expression that includes the process PID.

So at the moment we are hardwiring the PID, which doesn't really make sense the test flaky (e.g. currently fails on GitHub CI)

# This diff
We make it so that the the test-process can retrieve information (in this case, the pid representation) from the debuggee.

Ideally, we'd have used a `peer:call()` to query it, but the peer implementation relies on a custom `user` that gets paused when the VM is paused.

Alternatively, we could have used an `erpc` call, but we purposely don't start distribution on the node running the test.

So instead, we use peer to create a hidden "inspector" node that we use to proxy rpc calls to the debuggee.

Reviewed By: TD5

Differential Revision: D77946948

fbshipit-source-id: 2c5d5c607a2102e2ab885055a2f1db5eb47c26bf