Avoid bogus matching in scope tests
Summary:
# Context
In the DAP scopes request suite we have a couple of places were we are using `?assertMatch()` for maps,  where we could be using `assertEqual()` instead.

# Problem
`assertMatch()` is typically not good for maps as it doesn't tell us anything about new keys that happen to be in the result

# This diff
Use `assertEqual()` instead. Interestingly, this already shows that we were currently not matching on all available keys

Reviewed By: TD5

Differential Revision: D78005769

fbshipit-source-id: 597dc3124d75c9b57b4bba14d73f6cd1ea1efaaf