Add 'reapply_breakpoints' to edb_server
Summary:
# Context
edb_server is the debugger process that runs on every debuggee node, and keeps track of every breakpoint.

# Problem
We currently have no way to reapply the breakpoints on a module once the module is reloaded.

# This diff
Adds a ```edb_server:reapply_breakpoints/1``` function that does this.
Adds ```edb_server_break:reapply_breakpoints/2```

Reviewed By: jcpetruzza

Differential Revision: D78100603

fbshipit-source-id: e6624186b95515dde947aa25ad1cb0412e729674