Add reapply_breakpoints/3 to edb_server_break
Summary:
# Context
We want to keep track of module substitutes as part of supporting debugging when `meck` is used. This would mean transferring breakpoints when a module is substituted.

# Problem
`reapply_breakpoints/2` allows us to re-set breakpoints on a module but does not allow us to choose a source and target module for this.

# This diff
Adds function `reapply_breakpoints/3` allowing us to specify a source and target module, and function `has_breakpoints/2` that will indicate whether a substitute already has breakpoints indicating something is wrong.

Reviewed By: TheGeorge

Differential Revision: D79091329

fbshipit-source-id: e5824e75d4d8f7ef3574805a0fa08e310f06b195