Add add_module_substitute to edb_server
Summary:
As part of supporting debugging when meck is used, when meck mocks a module, it uses a different substitute name for the module.

When users try to set breakpoints from VSCode, they will use the original module name, so something needs to translate this to the substitute (loaded) module.

Added function `add_module_substitute/2` to `edb_server` so that when trying to set a breakpoint on X, it will set it on it's substitute Y instead. If X had breakpoints set, they'll be transferred to Y, and this handles transitive substitutes as well.

Reviewed By: jcpetruzza

Differential Revision: D78744757

fbshipit-source-id: b2881ff5fad75d146f260301e73465540bee3745