Add remove_module_substitute to edb_server
Summary:
# Context
As part of supporting debugging when meck is used, when meck mocks a module, it uses a different substitute name for the module.

# Problem
When users try to set breakpoints from VSCode, they will use the original module name, so something needs to translate this to the substitute (loaded) module (D78744757) and then transfer breakpoints back to the original module when the substitute module is unloaded.

# This diff
Added function `remove_module_substitute/2` to `edb_server` so that all breakpoints on a substituted module will be transferred back to the original module when this function is called.

Reviewed By: jcpetruzza

Differential Revision: D78748644

fbshipit-source-id: a48b5611c0d3547ea33835a584c3fb53ff840f25