Allow edb_server to add multiple subscriptions at a time
Summary:
# Context
We want to allow `edb:subscribe()` to be called before attachment, but need to sync subcriptions in `node_monitor` to `edb_server`.

# Problem
To do this, we need to call `edb_events:subscribe/4` for each `{subscription(), pid()` pair stored in `node_monitor`'s data.

# This diff
Adds a clause to `edb_server:subscribe_to_events_impl` that allows for these lists, and generates monitors and updates the state accordingly.

Reviewed By: jcpetruzza

Differential Revision: D81041799

fbshipit-source-id: ae2f395df30a0e7a371dfedd91d1a99784a1e7db