Allow edb:subscribe() to be called before attachment
Summary:
# Context
As part of multi-node debugging support, we need to issue a `startDebugging` reverse-request for each new node (T227782390).

# Problem
This is tricky to do with Reference notification messages, and using `edb:event()` which is used everywhere else in edb, would be much cleaner, however, we can't subscribe to events before reverse attaching as of now, which defeats the purpose.

# This diff
Modifies `edb_node_monitor:subscribe_to_events_impl` to allow subscribing to events even when we are not attached, by adding local subcriptions to `edb_server` during bootstrapping.

Reviewed By: jcpetruzza

Differential Revision: D80464903

fbshipit-source-id: 8d9c300a9979f7e41d18d4406af254720c0fea35