Add reverse_attach_timeout event type to edb_event()
Summary:
# Context
As part of multi-node debugging support, we need to issue a `startDebugging` reverse-request for each new node that attempts to attach.

# Problem
We receive an event when `edb:reverse_attach()` is called, but we don't have proper event handling that distinguishes between timeout scenarios (which can only happen for the first node) and regular attachment scenarios (first and additional nodes).

# This diff
Seperates the `reverse_attach` event into:
*   `{reverse_attach_timeout, reference()}` for timeout scenarios (first node only, since additional nodes can't timeout)
*   `{reverse_attach, reference(), reverse_attachment_event()}` for all attachment scenarios (first and additional nodes, both success and failure cases)

Reviewed By: jcpetruzza

Differential Revision: D80096279

fbshipit-source-id: 8e2b21d975754877678892423df8cdc84636521f