Change reverse_attach notifications to edb events
Summary:
# Context

As part of multi-node debugging support, we need to issue a `startDebugging` reverse-request for each new node (T227782390).

# Problem

This is tricky to do with Reference notification messages, and using `edb:event()` which is used everywhere else in edb, would be much cleaner.

# This diff

Removes `notification_reference` and `caller_pid` from `edb_node_monitor` by replacing it's usage in `reverse_attach_impl()` with a new `edb:event()`: `reverse_attach`, and encapsulates the reference in this event. Refactors other edb modules to handle this change, and refactors dependent test suites for the same.

Reviewed By: jcpetruzza

Differential Revision: D80087634

fbshipit-source-id: da3ed26a8b9c0a491d46421292a369b93c8836d6