Broadcast '{reverse_attach, reference(), {attached, node()}}' event from edb_node_monitor
Summary:
# Context
As part of multi-node debugging support, we need to issue a startDebugging reverse-request for each new node that attempts to attach.

# Problem
We receive an event when `edb:reverse_attach()` is called, but we don't do anything when a subsequent reverse attach request is made.

# This diff
Adds a function clause for `edb_node_monitor:reverse_attach_notification_impl` that broadcasts `{reverse_attach, reference(), {attached, node()}}` for this case. D80168956 gates this feature.

Reviewed By: jcpetruzza

Differential Revision: D80097569

fbshipit-source-id: 764a2383af01b682c65e772ac0e01ff869a44fe1