Add setup_multi_node_support to edb_bootstrap
Summary:
# Context
As part of multi-node debugging support, we need to issue a startDebugging reverse-request for each new node that attempts to attach.

# Problem
We receive an event when `edb:reverse_attach()` is called, but we don't do anything to steup multi-node debugging when a child node is created.

# This diff
Adds `setup_multi_node_support` to `edb_bootstrap` that sets `ERL_AFLAGS` to be code that will be evaluated when child nodes are created, automatically triggering a reverse attach.

Reviewed By: jcpetruzza

Differential Revision: D80610282

fbshipit-source-id: 80f4823fbbd871a8e5eb1017583f649158e216c4