Ensure stdin/stdout of DAP server are in binary mode (#11)
Summary:
Pull Request resolved: https://github.com/WhatsApp/edb/pull/11

# Context
The DAP server communicates with the client via a pipe on stdin/stdout. The protocol uses `\r\n` to delimit frames

# Problem
By default stdin/stdout are in text mode, and when we create a port in `edb_dap_transport`, even though we specify `binary` for the port, the underlying devices are already in `text` mode, so all `\r` are converted by default to `\n`.

This then means that the parsing of DAP frames fail, as we don't get the `\r\n\r\n` pattern that we expect

# This diff
Set stdio/stdout to binary mode before opening the port

Reviewed By: TheGeorge, robertoaloi

Differential Revision: D81579173

fbshipit-source-id: 4ff329ad2a01e6fa96db24d66147373572420808