Add option enable_multi_node to edb:reverse_attach()
Summary:
# Context
As part of multi-node debugging support, we need to issue a startDebugging reverse-request for each new node that attempts to attach.

# Problem
We need to make this an optional part of reverse attaching, since we want the experience of debugging and reverse attaching onto one node to remain the same.

# This diff
Adds `multi_node_enabled` and `add_flags_necessary` as `boolean()` options to `edb:reverse_attach/1`. The `node_attached` event is now only sent when `multi_node_enabled` is set to `true`. It is by default `false`. `add_flags_necessary` is true by default and specifies if the user requires edb to setup ERL_AFLAGS to inject reverse attach code, or if they have already done so.

Reviewed By: jcpetruzza

Differential Revision: D80168956

fbshipit-source-id: 003f868f9ddb1f38d89cf5a6414aa919bd2a3810