Update toolchain to rebased version
Summary:
# Context
Latest edb uses its own version of OTP, with some patches that still haven't been merged.

# Problem
We want to add a couple of new patches, but these are based on a newer version of upstream. So as a first step, we would like to update the current version we are using, to one that is rebased to a recent commit, so it is then easier to diff changes

# This diff
Update the toolchain.

Old version: https://github.com/jcpetruzza/otp/commits/edb-2025-05-08/
New version:
New version: https://github.com/jcpetruzza/otp/commits/edb-2025-09-12/

Reviewed By: TD5

Differential Revision: D82453617

fbshipit-source-id: 68f03684cee1957260e49e0b93762ecbfd364c2a