Add nodes/0 function to edb
Summary:
# Context
As part of multi-node debugging support, we need to issue a startDebugging reverse-request for each new node that attempts to reverse attach.

# Problem
We don't have anything in the `edb` API to access the additional nodes that connect as of now, like how we have `edb:attached_node()` for the main debugee node.

# This diff
Adds `edb:additional_nodes()` that returns a map of additional nodes, `#{node() => []}.

Reviewed By: michalmuskala

Differential Revision: D80611768

fbshipit-source-id: 4773b08d60f695dd38c59de67ff6a779dbffe09e

Co-authored-by: Agastya Bahl <agastyabahl@meta.com>