Create reverse_request_start_debugging.erl with startDebugging request
Summary:
# Context
As part of multi-node debugging support, we need to issue a startDebugging reverse-request for each new node that attempts to reverse attach. We currently have an event system where `edb_dap_internal_events` receives a `reverse_attach` event when any node, child or parent attempts to reverse attach.

# Problem
When we receive this event in DAP, we need to issue a `startDebugging` request and need a module that creates this for us, similar to the `runInTerminal` reverse request.

# This diff
Adds `reverse_request_start_debugging.erl`, that issues a `startDebugging` request if passed with `reverse_request_start_debugging:arguments()`.

Reviewed By: michalmuskala

Differential Revision: D80798353

fbshipit-source-id: 5a234011bc9a1bca9edb1268abb5c9cd8ff4b32a

Co-authored-by: Agastya Bahl <agastyabahl@meta.com>