Add reverse_attach_ref field to DAP server types
Summary:
# Context
When we receive a `reverse_attach` event in DAP, we check if the ref is the same as the one stored in the DAP server state.

# Problem
We do not have a field for `reverse_attach_ref` in the configuring and attached states.

# This diff
Adds a field `reverse_attach_ref` to these states that are `reference() | undefined`

Reviewed By: michalmuskala

Differential Revision: D81452980

fbshipit-source-id: c7e73454cddb938907e7b61f4ca69f62297cc22c

Co-authored-by: Agastya Bahl <agastyabahl@meta.com>