Make binary values structured
Summary:
# Context
When displaying variables, evaluation results, etc, values that have structured are shown truncated with ellipsis, and can then be expanded to show subfields, etc.

# Problem
Binary values are currently treated as an atomic value and have no structure. That means that it is not possible to observe the rest of the values. Moreover, because we only provide an `evaluateName` only for structured values, it is also not possible to copy it to the clipboard completely.

# This diff
Make binary values expandable. Notice that `binary:at()` indexes from 0, not 1, so unlike lists, we number elements from 0.

Reviewed By: michalmuskala

Differential Revision: D84987599

fbshipit-source-id: 0ba7e4d32b8c2b50372a17c3afe5ef161d8602ae