Fix flakiness in edb_dap_session_SUITE
Summary:
# Context
Some test in `edb_dap_session_SUITE` check that after sending a `disconnect` request, the debuggee node is killed. For this, we rely on `peer:get_state()`, and check that after the request completes, the peer is down.

# Problem
On the happy path, we kill the debuggee node by doing an rpc call that runs `erlang:halt()`, and the moment the `rpc:call()` is aborted, we consider it down.

However, it may take some extra time for the node to go down, or for the local peer to discover it, so the subsequent call to `peer:get_state()` may return `running`, which makes the test fail

# This diff
We now poll, for 5s, and only fail if `peer` never reports that the node is down.

Reviewed By: michalmuskala

Differential Revision: D85446890

fbshipit-source-id: 4fd9e4f748efdad4bcbc42d8727d6e52b210a311