[dap[ Sort thread list by PID
Summary:
# Context
The DAP specification has a notion of `thread` that in Erlang corresponds to a running process. In particular the `threads` request lets the DAP client (ie, the IDE), get the list of all processes in the current node.

# Problem
We are currently returning a list of process without any order. This means that if the user knows a pid, they can't leverage the ordering to quickly find the process in the "threads" view ("Call Stack" in VSCode).

This is compounded by the fact that VSCode currently doesn;t offer good support for searching threads when the number of threads is high, so finding a pid becomes extremely hard

# This diff
Sort the list of processes by pid (tbh, I'm not sure why we weren't doing this before)

Reviewed By: TD5

Differential Revision: D86027386

fbshipit-source-id: 7eb194fc078aae3a095caf25fd0a28ea89da50ce