Bump toolchain to include bif-timer pausing
Summary:
# Context
RIght now, when suspending a process, the VM also suspends the "process timer", that is, the timer associated with a running `receive` statement.

# Problem
Timers associated to `erlang:send_after()` or `erlang:start_timer()` BIFs are currently not paused, which breaks the "stop-the-world" illusion

# This diff
Bump the OTP toolchain to include patches by to also stop these timers. Notice that starting a timer on a suspended process will create a paused timer.

Diff of changes [here](https://github.com/jcpetruzza/otp/compare/edb-2025-09-25..edb-2025-11-13)

Reviewed By: TD5, michalmuskala

Differential Revision: D86968421

fbshipit-source-id: 7a041a4286496e4f8ee43ec0ae0660513efa0865