Improve error-handling when suspending/resuming processes
Summary:
# Context
When a process hits a breakpoint, edb suspends all processes. It does this by getting the list of processes and then calling `erlang:suspend_process()` on each. Of course, it can happen that between the call to `erlang:processes()` and `edb` trying to suspend a process, the process has already exited, in which case the call to `erlang:suspend_process()` will fail, which we want to handle gracefully by ignoring it. When resuming, the situation is similar.

# Problem
There was a typo where we'd try to catch `bardarg` instead of `badarg`.

Morever, the current logic only works on `error`, while in principle it could also happen that the call does an `exit(badarg)`, so let's handle that as well

# This diff
Generalize the logic, to handle either `error` or `exit`.

Reviewed By: michalmuskala

Differential Revision: D86968422

fbshipit-source-id: 14ece4ab7db6119964342f1b534a69367fc0982d