Remove redundant subscription to events
Summary:
# Context
The dap server needs to call `edb:subscribe()` to get notified when a process hits a breakpoint and other types of events.

# Problem
It used to be the case we could only call `edb:subscribe()` once we were attached, so we'd do this only after the DAP configuration phase was done, at which point all breakpoints have been set, etc.

However as part of the work to support debugging multiple-nodes, we recently generalized the API so that `edb:subscribe()` can be called before a reverse-attach, so that we can use the event mechanism to notify when a launched node reverse-attached.

The problem is that while we made it mandatory for a subscription to be already present in the `configured` state, we didn't remove the code that was subscribing when we reached that state. That meant that we'd end up with a second redundant subscription, and warnings in logs of the form:

> Invalid Subscription, skipping

# This diff
We can just remove that second subscription.

Reviewed By: michalmuskala

Differential Revision: D86972695

fbshipit-source-id: f97af70933d869f61f0e7a1ba2ec793ca1e62110