Ensure evaluating injected code blocks further -eval args
Summary:
# Context
When launching a new debuggee node, we generate a snippet of Erlang code that needs to be injected somhow via an `-eval` argument when executing the `erl` command. This code is expected to run first, and contacts back the debugger, which will then bootstrap `edb`, by loading all relevant modules, starting the `edb_server` and pausing the VM.

The expectation was always that in:

```
$ erl -eval EDB_INJECTED_CODE -eval USER_CODE
```

`USER_CODE` would not start running until `edb:continue()` is executed on the debugger node, as this let's the debugger set breakpoints, etc, without racing against user code.

# Problem
The injected code connects to the debugger node and then essentially does:

```
ok = gen_server:call(....)
````

This call triggers the bootstrapping process and was expected to block until  the `edb:continue()` is executed.

However, under the hood, this call was triggering the bootstapping process using a "cast", not a "call", so effectively the `call` would finish immediately and the next `-eval` block would start right away.

# This diff
Use a `call` instead of a `cast` to start the bootstrapping

Reviewed By: michalmuskala

Differential Revision: D87856204

fbshipit-source-id: 4de58f74b168aedfbb719a512bbad21b0c4ff49e