Avoid dead-lock loading module when setting a breakpoint
Summary:
# Context
When a node runs in `interactive` mode, modules are loaded as needed. This has implications for setting breakpoints (explicit or for supporting steps): the module may not be yet loaded, so we need to load it first. For this we use `code:ensure_loaded(Mod)`

# Problem
There's currently a potential dead-lock when we try to set a breakpoint on a paused VM. To see why, we need to look into how the code-loader works.

* `code:ensure_loaded(Mod)` involves codes to the `code_server`, but has no timeout
* The actual loading of a module is a dance between the code-server and the process that wants to load a module. In particular, the code-server first determines the code of the module to load and gives it back to the module to do the loading. At this point, the module is in a "waiting" state, and any other attempt to ensure the module is loaded, will add the process doing the call to a waiting-list, to be notified when the first process finishes

So the tricky scenario is this:
* Process P calls `code:ensure_loaded(my_mod)`; the code-server handles it the code to load and it now needs to load it.
* While this happens, the VM is paused by `edb_server`, in particular P gets suspended.
* The user attempts to set a breakpoint on `my_mod`, so `edb_server` calls `code:ensure_loaded(my_mod)`.
* Because `my_mod` is loading, the `edb_server process` is put in the waiting-list, but because P was suspended, it will never load. Deadlock!

# This diff
We address this by running `code:ensure_loaded()` inside a separate process and timing out if it doesn't respond fast enough. There is a very small probability that after hitting breakpoint there is some module on which we can't set breakpoints, but an error message is better than hanging.

Reviewed By: robertoaloi

Differential Revision: D87868407

fbshipit-source-id: 4759a8f764fb8f7608064174fc5696cf0c3851b9