Fix memory leak in book-keeping of breakpoints
Summary:
# Context
The `edb_server_break:breakpoints()` type is used for book-keeping on the breakpoints set in the VM. In particular, `vm_breakpoints` holds, for each breakpoints set, the reasons why we set it: due to an explicit user request and/or for the sake of stepping one or more processes.

Conceptually , `vm_breakpoints` is a:

```
#{{module(), line()} => [vm_breakpoint_reason()]}
```

However in practice, we want to be able to index by module, and avoid duplicate reasons, so we use a "curried" form:

```
#{module() => #{line()} => #{vm_breakpoint_reason() => []}}}
```

# Problem
When removing a breakpoint from the VM, we need to update the book-keeping and, in particular, remove the entry from `vm_breakpoints`. When no reasons remain, we need to remove the module-line pair.

However, we were doing this via a call to `maps:remove({Module, Line}, VmBreakpoints)`, but as discussed above, no such key exists in that map., thus the leak.

Because currently we have a redundant representation (being removed later in this stack), the leak wouldn't be observed, as we'd use another field as the primary source of truth.

# This diff
Add to `edb_server_maps` auxiliary functions to remove entries from a curried maps, and use that instead

Reviewed By: TheGeorge

Differential Revision: D88045710

fbshipit-source-id: 5418209b12216de22768473de3dc26976e7d2d69