Fix error in book-keeping of breakpoints
Summary:
# Context
The `edb_server_break:breakpoints()` type is used for book-keeping on the breakpoints set in the VM. In particular, `vm_breakpoints` holds, for each breakpoints set, the reasons why we set it: due to an explicit user request and/or for the sake of stepping one or more processes.

Conceptually , `vm_breakpoints` is a:

```
#{{module(), line()} => [vm_breakpoint_reason()]}
```

However in practice, we want to be able to index by module, and avoid duplicate reasons, so we use a "curried" form:

```
#{module() => #{line()} => #{vm_breakpoint_reason() => []}}}
```

# Problem
When a process has finished "stepping", we need to remove all the breakpoints that were introduced **only** for the sake of this process stepping.

The logic for doing this was flawed, though, and we'd remove from `vm_breakpoints` the information for any other line of that module.

The impact of this was that later other breakpoints would not be actually removed from the VM for that module, but because we were not using `vm_breakpoints` as the primary source-of-truth to decide if a breakpoints should be taken, we wouldn't notice.

# This diff
Fix the update of `vm_breakpoints` so that only one line is affected

Reviewed By: TheGeorge

Differential Revision: D88047138

fbshipit-source-id: fcdc85a0cc74a822c134ed18eb8774c72e44acf3