Remove redundancy from book-keeping of breakpoints
Summary:
# Context
The `edb_server_break:breakpoints()` type is used for book-keeping on the breakpoints set in the VM. One of tields we currently keep there `explicits` that gives us, conceptually, all the lines of all the modules where we've set an "explicit" breakpoints, that is, one expliictly requested by the user (and not one for stepping)

# Problem
This information is already readily available in the `vm_breakpoints` field, we we keep, for every line with a breakpoints, what are the reasons for setting it.

The redundancy is bad because:

- Adds a lot of complexity to the code in order to maintain an invariant that doesn't add much value.
- As we are trying to add breakpoints on functions, the complexity will become evern wors.
- It was already masking the fact that the invariant was broken, as witnessed by the previous bugs fixed in this stack (D88045710, D88047138).

# This diff
Remove that field and rely only on `vm_breakpoints`. We now how to pay a small conversion to resolve the `get_explicits()` call, but this is in practice a seldomly used function

Reviewed By: TheGeorge

Differential Revision: D88047252

fbshipit-source-id: ed004f5d4de3f35eb271e86bd3ceea4f523b85b4